NXT-653 Rename InfogramEmbed and related components to start with Storyblok for consistency
