+++ web/src/components/Storyblok/StoryblokInfogramEmbed/ClientInfogramEmbed.tsx
@@ -4,7 +4,7 @@ const LoadingPlaceholder = () => <div>Loading Infogram...</div>;
 
 export const ClientInfogramEmbed = dynamic(
 	/* webpackChunkName: "infogram-embed" */
-	() => import("./InfogramEmbed").then((mod) => mod.InfogramEmbed),
+	() => import("./StoryblokInfogramEmbed").then((mod) => mod.InfogramEmbed),
 	{
 		ssr: false,
 		loading: () => <LoadingPlaceholder />,
