+++ web/src/components/Storyblok/StoryblokInfogramEmbed/StoryblokInfogramEmbed.test.tsx
@@ -3,7 +3,7 @@ import { ScriptProps } from "next/script";
 
 import { InfogramEmbedStoryblok } from "@/types/storyblok";
 
-import { InfogramEmbed } from "./InfogramEmbed";
+import { StoryblokInfogramEmbed } from "./StoryblokInfogramEmbed";
 
 // Mock next/script to simulate script loading
 jest.mock("next/script", () => (props: ScriptProps) => {
@@ -35,7 +35,7 @@ describe("InfogramEmbed", () => {
 	});
 
 	it("renders with valid URL and default variants", () => {
-		render(<InfogramEmbed blok={mockInfogramProps} />);
+		render(<StoryblokInfogramEmbed blok={mockInfogramProps} />);
 		const embed = screen.getByTestId(infogramId);
 		expect(embed).toBeInTheDocument();
 		expect(embed).toHaveAttribute("data-id", infogramId);
@@ -44,7 +44,7 @@ describe("InfogramEmbed", () => {
 	});
 	it("renders with RichText", () => {
 		render(
-			<InfogramEmbed
+			<StoryblokInfogramEmbed
 				blok={{
 					...mockInfogramProps,
 					displayMode: "withRichText",
@@ -56,10 +56,10 @@ describe("InfogramEmbed", () => {
 	});
 	it("renders with a full width layoutVariant", () => {
 		render(
-			<InfogramEmbed
+			<StoryblokInfogramEmbed
 				blok={{
 					...mockInfogramProps,
-					layoutVariant: "fullwidth",
+				layoutVariant: "fullwidth",
 				}}
 			/>
 		);
@@ -70,7 +70,7 @@ describe("InfogramEmbed", () => {
 
 	it("shows error message when URL is invalid", () => {
 		render(
-			<InfogramEmbed
+			<StoryblokInfogramEmbed
 				blok={{
 					...mockInfogramProps,
 					infogramUrl: "",
@@ -89,7 +89,7 @@ describe("InfogramEmbed", () => {
 		document.body.appendChild(script);
 
 		render(
-			<InfogramEmbed
+			<StoryblokInfogramEmbed
 				blok={{
 					...mockInfogramProps,
 					infogramUrl: validUrl,
@@ -107,7 +107,7 @@ describe("InfogramEmbed", () => {
 		document.body.appendChild(script);
 
 		render(
-			<InfogramEmbed
+			<StoryblokInfogramEmbed
 				blok={{
 					...mockInfogramProps,
 					infogramUrl: validUrl,
